requirements.txt
discord.py==1.7.3
python-dotenv==0.19.1
python-mpd2==3.0.4